backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["only-username-available <santiago.morillo@wizeline.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
boto3 = "^1.35.70"
openai = "^1.55.2"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-multipart = "^0.0.17"
langchain-aws = "^0.2.7"
langchain = "^0.3.16"
langfuse = "^2.58.0"
sse-starlette = "^2.1.3"
langchain-groq = "^0.2.3"
langchain-openai = "^0.3.2"
rapidfuzz = "^3.11.0"
langchain-community = "^0.3.16"
faiss-cpu = "^1.9.0.post1"
numpy = "^2.2.2"
pandas = "^2.2.3"
openpyxl = "^3.1.5"
thefuzz = "^0.22.1"
xmltodict = "^0.14.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
